made rollback for sanity check, `csv_to_database` functionality mover to `csv_importer`, fixed imports, changed mocks
